refactor: moved all unit tests to tests dir
