+++ tests/db_test.go
@@ -1,10 +1,11 @@
-package repository
+package tests
 
 import (
 	"fmt"
 	"os"
 	"testing"
 
+	"github.com/codescalersinternships/Link-Tree-Dohaelsawy/backend/database/repository"
 	model "github.com/codescalersinternships/Link-Tree-Dohaelsawy/backend/models"
 	"github.com/stretchr/testify/suite"
 	"gorm.io/driver/postgres"
@@ -13,7 +14,7 @@ import (
 
 type DatabaseTestSuite struct {
 	suite.Suite
-	dbInstance DbInstance
+	DbInstance repository.DbInstance
 }
 
 func TestSuite(t *testing.T) {
@@ -30,26 +31,26 @@ func (suite *DatabaseTestSuite) SetupSuite() {
 	db, err := gorm.Open(postgres.Open(conString), &gorm.Config{})
 	suite.Require().NoError(err, "Error connecting to the test database")
 
-	suite.dbInstance.DB = db.Debug()
+	suite.DbInstance.DB = db.Debug()
 
-	err = suite.dbInstance.DB.AutoMigrate(&model.User{}, &model.Link{})
+	err = suite.DbInstance.DB.AutoMigrate(&model.User{}, &model.Link{})
 	suite.Require().NoError(err, "Error auto-migrating database tables")
 
 	user := model.User{FirstName: "doha", LastName: "elsawy", Email: "doha@gmail.com", Password: "password", Username: "do23"}
-	err = suite.dbInstance.AddNewUser(&user)
+	err = suite.DbInstance.AddNewUser(&user)
 	suite.Require().NoError(err, "Error adding user before testing")
 
 	link := model.Link{Name: "youtube", Url: "youtube.com", UserID: 1}
-	err = suite.dbInstance.AddNewLink(&link)
+	err = suite.DbInstance.AddNewLink(&link)
 	suite.Require().NoError(err, "Error creating link record")
 }
 
 func (suite *DatabaseTestSuite) TearDownSuite() {
 
-	err := suite.dbInstance.DB.Exec("DROP TABLE links;").Error
+	err := suite.DbInstance.DB.Exec("DROP TABLE links;").Error
 	suite.Require().NoError(err, "Error dropping test table")
 
-	err = suite.dbInstance.DB.Exec("DROP TABLE users;").Error
+	err = suite.DbInstance.DB.Exec("DROP TABLE users;").Error
 	suite.Require().NoError(err, "Error dropping test table")
 
 }
