+++ tests/linkQueries_test.go
@@ -1,4 +1,4 @@
-package repository
+package tests
 
 import (
 	model "github.com/codescalersinternships/Link-Tree-Dohaelsawy/backend/models"
@@ -8,11 +8,11 @@ func (suite *DatabaseTestSuite) TestLinkAddNewLink() {
 
 	link := model.Link{Name: "youtube", Url: "youtube.com", UserID: 1}
 
-	err := suite.dbInstance.AddNewLink(&link)
+	err := suite.DbInstance.AddNewLink(&link)
 	suite.Require().NoError(err, "Error creating link record")
 
 	var retrievedLink model.Link
-	err = suite.dbInstance.GetOneLink(&retrievedLink, link.ID)
+	err = suite.DbInstance.GetOneLink(&retrievedLink, link.ID)
 	suite.Require().NoError(err, "Error retrieving link record")
 
 	suite.Require().Equal(link.Name, retrievedLink.Name)
@@ -23,11 +23,11 @@ func (suite *DatabaseTestSuite) TestLinkDeleteLink() {
 
 	link := model.Link{ID: 1, Name: "youtube", Url: "youtube.com", UserID: 1}
 
-	err := suite.dbInstance.DeleteLink(&link, link.ID)
+	err := suite.DbInstance.DeleteLink(&link, link.ID)
 	suite.Require().NoError(err, "Error deleting link record")
 
 	var retrievedLink model.Link
-	err = suite.dbInstance.GetOneLink(&retrievedLink, link.ID)
+	err = suite.DbInstance.GetOneLink(&retrievedLink, link.ID)
 	suite.Require().Error(err, "Error retrieving link record")
 }
 
@@ -35,12 +35,12 @@ func (suite *DatabaseTestSuite) TestLinkPutOneLink() {
 
 	link := model.Link{Name: "instagram", Url: "instagram.com", UserID: 1}
 
-	err := suite.dbInstance.AddNewLink(&link)
+	err := suite.DbInstance.AddNewLink(&link)
 	suite.Require().NoError(err, "Error add link record")
 
 	link.Name = "x"
 
-	err = suite.dbInstance.PutOneLink(&link, link.ID)
+	err = suite.DbInstance.PutOneLink(&link, link.ID)
 	suite.Require().NoError(err, "Error update link record")
 	suite.Require().Equal("x", link.Name)
 }
@@ -48,7 +48,7 @@ func (suite *DatabaseTestSuite) TestLinkPutOneLink() {
 func (suite *DatabaseTestSuite) TestLinkGetAllLinksForUser() {
 	var links []model.Link
 
-	err := suite.dbInstance.GetAllLinksForUser(&links, 1)
+	err := suite.DbInstance.GetAllLinksForUser(&links, 1)
 
 	suite.Require().NoError(err, "Error retrieving links for specified user record")
 	suite.Require().NotZero(links)
