+++ tests/userQueries_test.go
@@ -1,4 +1,4 @@
-package repository
+package tests
 
 import (
 	model "github.com/codescalersinternships/Link-Tree-Dohaelsawy/backend/models"
@@ -8,11 +8,11 @@ func (suite *DatabaseTestSuite) TestLinkAddNewUser() {
 
 	user := model.User{FirstName: "doha", LastName: "elsawy", Email: "x@gmail.com", Password: "password", Username: "do2333"}
 
-	err := suite.dbInstance.AddNewUser(&user)
+	err := suite.DbInstance.AddNewUser(&user)
 	suite.Require().NoError(err, "Error adding user before testing")
 
 	var retrievedUser model.User
-	err = suite.dbInstance.GetUserEmail(&retrievedUser, user.Email)
+	err = suite.DbInstance.GetUserEmail(&retrievedUser, user.Email)
 	suite.Require().NoError(err, "Error retrieving link record")
 	suite.Require().Equal(user.Email, retrievedUser.Email)
 	suite.Require().Equal(user.Password, retrievedUser.Password)
@@ -22,11 +22,11 @@ func (suite *DatabaseTestSuite) TestLinkDeleteUser() {
 
 	user := model.User{FirstName: "doha", LastName: "elsawy", Email: "c@gmail.com", Password: "password", Username: "dosfs"}
 
-	err := suite.dbInstance.DeleteUser(&user, user.ID)
+	err := suite.DbInstance.DeleteUser(&user, user.ID)
 	suite.Require().NoError(err, "Error deleting link record")
 
 	var retrievedUser model.User
-	err = suite.dbInstance.GetUserID(&retrievedUser, user.ID)
+	err = suite.DbInstance.GetUserID(&retrievedUser, user.ID)
 	suite.Require().Error(err, "Error retrieving link record")
 }
 
@@ -34,12 +34,12 @@ func (suite *DatabaseTestSuite) TestLinkPutOneUser() {
 
 	user := model.User{FirstName: "doha", LastName: "elsawy", Email: "c@gmail.com", Password: "password", Username: "dosfs"}
 
-	err := suite.dbInstance.AddNewUser(&user)
+	err := suite.DbInstance.AddNewUser(&user)
 	suite.Require().NoError(err, "Error adding user before testing")
 
 	user.FirstName = "soha"
 
-	err = suite.dbInstance.PutOneUser(&user, user.ID)
+	err = suite.DbInstance.PutOneUser(&user, user.ID)
 	suite.Require().NoError(err, "Error update link record")
 	suite.Require().Equal("soha", user.FirstName)
 }
